Additional refactoring: * UnsupportedSystemException extends CommandFailedException * UnsupportedSystemException is now a static inner class of Mounter interface * added some javadoc
