+++ src/main/java/org/cryptomator/frontend/webdav/mount/Mounter.java
@@ -1,7 +1,6 @@
 package org.cryptomator.frontend.webdav.mount;
 
 import java.net.URI;
-import java.net.URL;
 import java.nio.file.Path;
 import java.util.Optional;
 
@@ -46,6 +45,9 @@ public interface Mounter {
 
 	}
 
+	/**
+	 * Thrown if the mount or unmount operation for a webdav directory fails.
+	 */
 	class CommandFailedException extends Exception {
 
 		public CommandFailedException(String message) {
@@ -54,6 +56,27 @@ public interface Mounter {
 
 		public CommandFailedException(Throwable cause) {
 			super(cause);
+		}
+
+	}
+
+	/**
+	 * Thrown when the underlying OS is not supported mounting a webdav directory.
+	 * <p>
+	 * It contains the uri of the webdav service to offer callers the possiblilty of process it further.
+	 */
+	class UnsupportedSystemException extends CommandFailedException {
+
+		private static final String MESSAGE = "No applicable mounting strategy found for this system.";
+		private final URI uri;
+
+		public UnsupportedSystemException(URI uri) {
+			super(MESSAGE);
+			this.uri = uri;
+		}
+
+		public URI getUri() {
+			return uri;
 		}
 
 	}
